Refactor launcher and layout planning modules
