+++ src/windows_layout.rs
@@ -1,3 +1,6 @@
+pub mod apply;
+pub mod plan;
+
 use crate::plugins::layouts_storage::{
     Layout, LayoutMatch, LayoutPlacement, LayoutWindow, LayoutWindowState,
 };
